% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donothing.R
\name{donothing}
\alias{donothing}
\title{Do Nothing}
\usage{
donothing(arg)
}
\arguments{
\item{arg}{Anything.}
}
\value{
arg
}
\description{
That's that.
}
\examples{
tibble::tibble(x = 1:3) \%>\%
dplyr::mutate(y = letters[1:3]) \%>\%
# dplyr::mutate(y = LETTERS[1:3]) \%>\%
donothing()

tibble::tibble(x = 1:3) \%>\%
# dplyr::mutate(y = letters[1:3]) \%>\%
dplyr::mutate(y = LETTERS[1:3]) \%>\%
donothing()
}
